Splitted and improved ExportDataOptions.
